An idea for a simple testing framework
