+++ src/intset.h
@@ -50,8 +50,4 @@ uint32_t intsetLen(const intset *is);
 size_t intsetBlobLen(intset *is);
 int intsetValidateIntegrity(const unsigned char *is, size_t size, int deep);
 
-#ifdef REDIS_TEST
-int intsetTest(int argc, char *argv[], int flags);
-#endif
-
 #endif // __INTSET_H
