+++ src/quicklist.h
@@ -203,10 +203,6 @@ quicklistNode *quicklistBookmarkFind(quicklist *ql, const char *name);
 void quicklistBookmarksClear(quicklist *ql);
 int quicklistisSetPackedThreshold(size_t sz);
 
-#ifdef REDIS_TEST
-int quicklistTest(int argc, char *argv[], int flags);
-#endif
-
 /* Directions for iterators */
 #define AL_START_HEAD 0
 #define AL_START_TAIL 1
