+++ src/sds.h
@@ -280,8 +280,4 @@ void *sds_malloc(size_t size);
 void *sds_realloc(void *ptr, size_t size);
 void sds_free(void *ptr);
 
-#ifdef REDIS_TEST
-int sdsTest(int argc, char *argv[], int flags);
-#endif
-
 #endif
