+++ src/server.c
@@ -6920,91 +6920,10 @@ int iAmMaster(void) {
             (server.cluster_enabled && nodeIsMaster(server.cluster->myself)));
 }
 
-#ifdef REDIS_TEST
-#include "testhelp.h"
-#include "intset.h"  /* Compact integer set structure */
-
-int __failed_tests = 0;
-int __test_num = 0;
-
-/* The flags are the following:
-* --accurate:     Runs tests with more iterations.
-* --large-memory: Enables tests that consume more than 100mb. */
-typedef int redisTestProc(int argc, char **argv, int flags);
-struct redisTest {
-    char *name;
-    redisTestProc *proc;
-    int failed;
-} redisTests[] = {
-    {"ziplist", ziplistTest},
-    {"quicklist", quicklistTest},
-    {"intset", intsetTest},
-    {"zipmap", zipmapTest},
-    {"sha1test", sha1Test},
-    {"util", utilTest},
-    {"endianconv", endianconvTest},
-    {"crc64", crc64Test},
-    {"zmalloc", zmalloc_test},
-    {"sds", sdsTest},
-    {"dict", dictTest},
-    {"listpack", listpackTest}
-};
-redisTestProc *getTestProcByName(const char *name) {
-    int numtests = sizeof(redisTests)/sizeof(struct redisTest);
-    for (int j = 0; j < numtests; j++) {
-        if (!strcasecmp(name,redisTests[j].name)) {
-            return redisTests[j].proc;
-        }
-    }
-    return NULL;
-}
-#endif
-
 int main(int argc, char **argv) {
     struct timeval tv;
     int j;
     char config_from_stdin = 0;
-
-#ifdef REDIS_TEST
-    if (argc >= 3 && !strcasecmp(argv[1], "test")) {
-        int flags = 0;
-        for (j = 3; j < argc; j++) {
-            char *arg = argv[j];
-            if (!strcasecmp(arg, "--accurate")) flags |= REDIS_TEST_ACCURATE;
-            else if (!strcasecmp(arg, "--large-memory")) flags |= REDIS_TEST_LARGE_MEMORY;
-            else if (!strcasecmp(arg, "--valgrind")) flags |= REDIS_TEST_VALGRIND;
-        }
-
-        if (!strcasecmp(argv[2], "all")) {
-            int numtests = sizeof(redisTests)/sizeof(struct redisTest);
-            for (j = 0; j < numtests; j++) {
-                redisTests[j].failed = (redisTests[j].proc(argc,argv,flags) != 0);
-            }
-
-            /* Report tests result */
-            int failed_num = 0;
-            for (j = 0; j < numtests; j++) {
-                if (redisTests[j].failed) {
-                    failed_num++;
-                    printf("[failed] Test - %s\n", redisTests[j].name);
-                } else {
-                    printf("[ok] Test - %s\n", redisTests[j].name);
-                }
-            }
-
-            printf("%d tests, %d passed, %d failed\n", numtests,
-                   numtests-failed_num, failed_num);
-
-            return failed_num == 0 ? 0 : 1;
-        } else {
-            redisTestProc *proc = getTestProcByName(argv[2]);
-            if (!proc) return -1; /* test not found */
-            return proc(argc,argv,flags);
-        }
-
-        return 0;
-    }
-#endif
 
     /* We need to initialize our libraries, and the server configuration. */
 #ifdef INIT_SETPROCTITLE_REPLACEMENT
