+++ src/sha1.h
@@ -21,7 +21,4 @@ void SHA1Init(SHA1_CTX* context);
 __attribute__((noinline)) void SHA1Update(SHA1_CTX* context, const unsigned char* data, uint32_t len);
 void SHA1Final(unsigned char digest[20], SHA1_CTX* context);
 
-#ifdef REDIS_TEST
-int sha1Test(int argc, char **argv, int flags);
-#endif
 #endif
