+++ src/zipmap.h
@@ -47,8 +47,4 @@ size_t zipmapBlobLen(unsigned char *zm);
 void zipmapRepr(unsigned char *p);
 int zipmapValidateIntegrity(unsigned char *zm, size_t size, int deep);
 
-#ifdef REDIS_TEST
-int zipmapTest(int argc, char *argv[], int flags);
-#endif
-
 #endif
